requirements.txt
praw
fastapi
uvicorn
transformers
torch
python-dotenv
numpy